refactor: extract Event and field methods into event.go (no API change)
Third and final cut at the bolt.go split. Moves the Event type, the
event pool, and all 30+ field-builder methods into a focused event.go
(802 lines).

What moved out of bolt.go:
  * type Event struct and the global eventPool sync.Pool
  * (e *Event) Logger() — the conversion-back-to-Logger helper
  * The terminator: Msg, Send, Printf
  * The string family: Str, Stringer, Bytes, Hex, Base64, RandID
  * The numeric family: Int / Int8 / Int16 / Int32 / Int64,
    Uint / Uint8 / Uint16 / Uint32 / Uint64, Float64, Bool, Counter
  * The time family: Time, Dur, Timestamp
  * The IP / network: IPAddr
  * Composite types: Any, Interface, Fields, Ints, Strs, Dict
  * Diagnostics: Stack, Caller, CallerSkip, Err
  * Drops a long list of imports from bolt.go
    (crypto/rand, encoding/base64, encoding/hex, encoding/json, fmt,
    net, runtime, strings, sync, time) which were only used by the
    moved methods

What stayed in bolt.go:
  * Package doc-comment, constants
  * Level, Level.String, level constants and slog-style aliases
  * Handler interface, ErrorHandler, Hook + SampleHook
  * Logger type and (l *Logger) methods: New, SetErrorHandler,
    AddHook, With, Ctx, log, Info/Error/Debug/Warn/Trace/Fatal,
    SetLevel
  * Default logger, init, ParseLevel, isTerminal
  * Package-level Info/Error/etc. shortcuts
  * levelWriter + NewLevelWriter
  * exitFunc

Combined size after the three-cut split:
  * bolt.go:    508 lines (was 2002 originally)
  * event.go:   802 lines (new)
  * encode.go:  323 lines (extracted in #57)
  * handler.go: 352 lines (extracted in #55)

Also updates .golangci.yml to extend the eventPool.Get and rand.Read
errcheck exemptions to cover event.go (same pattern, just a different
file now).

Pure code-motion. No symbols added or removed; no public API change.

Verification:
  * `go test -short -count=1 ./...` — passes
  * `go test -race -short -count=1 .` — passes
  * `golangci-lint run` — 0 issues
  * `BenchmarkZeroAllocation` — 73 ns/op, 0 allocs/op (within budget)
