+++ .golangci.yml
@@ -102,7 +102,7 @@ linters:
         text: QF1001
       - linters:
           - errcheck
-        path: bolt.go
+        path: (bolt|event)\.go
         source: eventPool\.Get\(\)
       - linters:
           - errcheck
@@ -110,7 +110,7 @@ linters:
         source: bufPool\.Get\(\)
       - linters:
           - errcheck
-        path: bolt.go
+        path: (bolt|event)\.go
         source: rand\.Read\(id\)
       - linters:
           - revive
